requirements.txt
numpy==1.21.5
pandas==1.3.5
neomodel==4.0.8
Django==3.2.12
django-neomodel==0.0.7
djangorestframework==3.13.1
drf-renderer-xlsx==0.4.5
djangorestframework-csv==2.1.1
drf-yasg==1.20.0
django-material==1.10.0
django-grappelli==2.15.5
biopython==1.79
Whoosh==2.7.4
mysqlclient==2.1.0
greenlet==1.1.2
eventlet==0.30.2
gunicorn==20.1.0